Minor UI fixes
