+++ src/game/scenes/system/interface/building-info/controls/index.tsx
@@ -1,43 +1,18 @@
-import { useMobilePlatform } from 'phaser-react-ui';
 import React from 'react';
 
-import { Cost } from '~scene/system/interface/cost';
 import { BuildingControl } from '~type/world/entities/building';
 
-import {
-  Wrapper, Action, Label, Addon, Main, Key,
-} from './styles';
+import { Action } from './action';
+import { Wrapper } from './styles';
 
 type Props = {
   list: BuildingControl[]
 };
 
-export const BuildingControls: React.FC<Props> = ({ list }) => {
-  const isMobile = useMobilePlatform();
-
-  return (
-    <Wrapper>
-      {list.map((control) => (
-        <Action
-          key={control.label}
-          $disabled={control.disabled}
-          {...{
-            [isMobile ? 'onTouchEnd' : 'onClick']: control.onClick,
-          }}
-        >
-          <Main>
-            {!isMobile && (
-              <Key>{control.hotkey}</Key>
-            )}
-            <Label>{control.label}</Label>
-          </Main>
-          {!!control.cost && (
-            <Addon>
-              <Cost type="resources" value={control.cost} size="small" />
-            </Addon>
-          )}
-        </Action>
-      ))}
-    </Wrapper>
-  );
-};
+export const BuildingControls: React.FC<Props> = ({ list }) => (
+  <Wrapper>
+    {list.map((control) => (
+      <Action key={control.label} control={control} />
+    ))}
+  </Wrapper>
+);
